usersession: rename sessionagent package to usersession/agent
