+++ usersession/agent/response.go
@@ -17,7 +17,7 @@
  *
  */
 
-package sessionagent
+package agent
 
 import (
 	"encoding/json"
