+++ usersession/agent/session_agent.go
@@ -17,7 +17,7 @@
  *
  */
 
-package sessionagent
+package agent
 
 import (
 	"context"
